Edit AdminStatus and Login--Bugs
I have allowed the ability for the admin to change a users admin status.
However, it does not reflect on the actual Login.txt, but it will soon.

The Login section has a bug, ive had a hard time fixing it. Anytime the
user enters a WRONG login, it keeps that in the buffer and keeps using
it, I haven't been able to find a way around it yet.